requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openai==1.13.3
flask-cors==4.0.0
python-dotenv==1.0.1   # already present, keep
elevenlabs==1.4.1
